Clean-up annoying issues
Put a flag in place to indicate with the imu data notification is active or inactive.  Then used to flag to put the imu to sleep when notifications are inactive which stops the data transfer to the softdevice which was causing an annoying error message.

Put a flag in place to indicate when the imu data transfer is occurring.  I did this in an attempt to remove the occasional NRF_ERROR_RESOURCES error message.  But using this flag to indicate when the next transfer should occur only caused the data transfer to slow down to unusable data rate.  I left the code in place and commented out the use of the flag in characteristic_update_imu_data().  This returned the data rate to a usable amount but I still occationally get the NRF_ERROR_RESOURCES error.

Make some naming changes of functions.

Removed a couple of files which were not part of the project.