Restructure code
Restructured code by moving the support for the uart to it's own module.  This is in preparation of adding usb support instead of uart for just the pca10059 dongle.  The intent is to have the code structured so that the host can be either the uart for all devices or usb for those devices that support usb.