pca10059 usbd support
Added support for usb for the pca10059.  This replaces the uart host interface with a usb host interface.  This means that the pca10059 can be used as is comes out of the box without the need to solder on wires to get access to the uart pins needed for uart output.