Fixed release build
The release build is running faster than the debug build.  This was causing the output array of characters to be over-written with the next string prior to the prior string being fully outputted.  The fix was to copy the string to a static array that gets sent out.

Another fix would have been to create an array of strings and then index through the array with each output circling back to the first when the array reaches the end.