Interface Enhancements
Added the initialization of the array that holds the full scale resolutions for the accel, gyro, and mag by reading the values from the peripheral after the connection is made.

Added the ability to read a single imu data set.  But this only returns the last measurement that was made while the device was running; that is, only when notify is enabled.  The imu is downed down when notify is disabled to save power.

Corrected an annoying warning about const value.